+++ DX12Renderer/Source/Graphics/Backend/SwapChain.cpp
@@ -108,7 +108,7 @@ void SwapChain::CreateBackBufferTextures()
         D3D12_RESOURCE_DESC backBufferDesc = backBuffer->GetDesc();
 
         m_BackBuffers[i] = std::make_unique<Texture>("Back buffer", TextureDesc(TextureUsage::TEXTURE_USAGE_RENDER_TARGET,
-            TextureFormat::TEXTURE_FORMAT_RGBA8_UNORM, backBufferDesc.Width, backBufferDesc.Height));
+            TextureFormat::TEXTURE_FORMAT_RGBA8_UNORM, static_cast<uint32_t>(backBufferDesc.Width), static_cast<uint32_t>(backBufferDesc.Height)));
         m_BackBuffers[i]->SetD3D12Resource(backBuffer);
     }
 }
